requirements.txt
aiosignal==1.3.1
appdirs==1.4.4
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
cssselect==1.2.0
fake-useragent==1.4.0
frozenlist==1.4.1
greenlet==3.0.1
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
lxml==5.1.0
multidict==6.0.4
outcome==1.3.0.post0
packaging==23.2
parse==1.20.0
pillow==10.2.0
playwright==1.40.0
pyee==11.0.1
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selectolax==0.3.17
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
